+++ client/components/Admin/Admin.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import MenuBar from '../utils/MenuBar';
+
+const Admin = () => (
+	<div>
+		<MenuBar/>
+		<h1>You're an admin</h1>
+	</div>
+);
+
+export default Admin;
