+++ client/components/Admin/index.js
@@ -1,0 +1,1 @@
+export { default } from './Admin';
