+++ server/api/index.js
@@ -3,10 +3,10 @@ const v1 = require('./v1/v1API');
 
 const V1 = 'v1';
 
-module.exports = (ctr) => {
+module.exports = (controllers) => {
 	const api = Router();
 
-	api.use(`/${V1}/`, v1(ctr));
+	api.use(`/${V1}/`, v1(controllers));
 
 	return api;
 };
