+++ server/auth/index.js
@@ -40,7 +40,7 @@ function getBearer(req) {
  * @param {Object} req Incoming request to run through middleware.
  * @param {Object} res Response sent back if needed.
  * @param {Function} next Next middleware function.
- * @return {Function} Next middlware if authenticated, otherwise invalid response.
+ * @return {Function} Next middleware if authenticated, otherwise invalid response.
  */
 module.exports = () => (req, res, next) => {
