+++ server/controllers/admin.controller.js
@@ -1,0 +1,5 @@
+const { Admin } = require('../models');
+
+module.exports = {
+
+};
