+++ server/controllers/index.js
@@ -1,0 +1,9 @@
+const admin = require('./admin.controller');
+const auth = require('./auth.controller');
+const user = require('./user.controller');
+
+module.exports = {
+	admin,
+	auth,
+	user
+};
