+++ server/db/index.js
@@ -1,0 +1,3 @@
+const db = require('./db');
+
+module.exports = db;
